Subo cambios del tpa, y la estructura de los test, falta hacer un par

+++ tp-anual/src/main/java/ar/utn/frba/disenio/tp_anual/Rubro.java
@@ -1,13 +1,15 @@
 package ar.utn.frba.disenio.tp_anual;
 
+import java.math.BigDecimal;
+
 public class Rubro {
-	private Integer radioCercania;
+	private BigDecimal radioCercania;
 	private String nombreRubro;
-	public Rubro(String nombre, Integer radioCercania){
+	public Rubro(String nombre, BigDecimal radioCercania){
 		this.radioCercania = radioCercania;
 		this.nombreRubro = nombre;
 	}
-	public Integer getRadioCercania(){
+	public BigDecimal getRadioCercania(){
 		return this.radioCercania;
 	}
 	public String getNombre(){
